generator/generator_requirements.txt
# requirements
kafka-python